fix: object_store: stop throwing errors when checking generation file
A new routine is added to File class, to check whether a file exists or
not. It will allow file existence check, without trying to open the file
and thus throw an error when the use case is valid.

This new routine is added to Generation wasUpdated routine: if the
generation file does not exist, the routine will exit without error,
instead of trying to open the file and report an error.

Fixes #753

Signed-off-by: Alexandre Besnard <alexandre.besnard@softathome.com>